Explicit integer cast to avoid warning

See merge request molpro/sjef!126

+++ lib/sjef.cpp
@@ -800,7 +800,7 @@ void Project::clean(bool oldOutput, bool output, bool unused, int keep_run_direc
     throw std::invalid_argument("sjef::project::clean for unused files is not yet implemented");
   if (auto statuss = cached_status(); statuss == running || statuss == waiting)
     keep_run_directories = std::max(keep_run_directories, 1);
-  while (run_list().size() > keep_run_directories)
+  while (run_list().size() > size_t(keep_run_directories))
     run_delete(*run_list().rbegin());
 }
